requirements.txt
streamlit
pillow
opencv-python-headless
groq
python-dotenv
google-cloud-vision
